# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedineTech - ERP - Management Software API",
        "description": "Integration API Docs.",
        "contact": {
            "name": "Developers Team",
            "url": "https://medine.tech/",
            "email": "devs@medine.tech"
        },
        "version": "0.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Localhost Server"
        }
    ],
    "paths": {
...
        "/api/email/verification-notification": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send a new email verification notification",
                "operationId": "21ff3f9f9b9e1407909089f864312e1c",
                "responses": {
                    "200": {
                        "description": "Verification link sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "verification-link-sent"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}